Move getRange function to separate file
