+++ packages/react/src/Pagination/Pagination.tsx
@@ -7,57 +7,8 @@ import { ChevronLeftIcon, ChevronRightIcon } from '@amsterdam/design-system-reac
 import clsx from 'clsx'
 import { forwardRef } from 'react'
 import type { ForwardedRef, HTMLAttributes } from 'react'
+import { getRange } from './getRange'
 import { Icon } from '../Icon/Icon'
-
-/**
- * This returns an array of the range, including spacers
- *
- * @example
- * currentPage = 4, totalPages = 7, maxVisiblePages = 7
- * // returns [1, 2, 3, 4, 5, 6, 7]
- *
- * @example
- * currentPage = 5, totalPages = 100, maxVisiblePages = 7
- * // returns [1, 'firstSpacer', 4, 5, 6, 'lastSpacer', 100]
- *
- * @example
- * currentPage = 97, totalPages = 100, maxVisiblePages = 7
- * // returns [1, 'firstSpacer', 96, 97, 98, 99, 100]
- */
-
-function getRange(currentPage: number, totalPages: number, maxVisiblePages: number): Array<string | number> {
-  // the total amount of visible pages is whatever's lower, totalPages or maxVisiblePages
-  // maxVisiblePages has a lower limit of 5
-  const visiblePages = Math.min(totalPages, Math.max(maxVisiblePages, 5))
-
-  const min = 1
-  // the center part of the range starts with the current page minus half of the visible pages
-  let centerStartPage = currentPage - Math.floor(visiblePages / 2)
-  // centerStartPage has a lower limit of 1
-  centerStartPage = Math.max(centerStartPage, min)
-  // centerStartPage has an upper limit of 1 plus total pages minus visible pages
-  centerStartPage = Math.min(centerStartPage, min + totalPages - visiblePages)
-
-  const pages = Array.from({ length: visiblePages }, (_el, i) => centerStartPage + i).reduce<Array<string | number>>(
-    (acc, pageNr, index) => {
-      if (index === 0 && pageNr !== 1) {
-        return [1, 'firstSpacer']
-      }
-
-      if (totalPages > visiblePages && index === visiblePages - 2 && currentPage < totalPages - 2) {
-        return [...acc, 'lastSpacer', totalPages]
-      }
-      // Skip a number when spacer is already added
-      if ((acc.includes('firstSpacer') && index === 1) || (acc.includes('lastSpacer') && index === visiblePages - 1)) {
-        return acc
-      }
-      return [...acc, pageNr]
-    },
-    [],
-  )
-
-  return pages
-}
 
 export type PaginationProps = {
   /** The id of the component. */
